requirements.txt
discord.py
PyPDF2
dotenv